removed messagelist component.
